refactor(HeroOfferDetails): move getAddress function to an utils file
